ausoceantv: restructure dirs
