+++ cmd/ausoceantv/frontend/.prettierrc
