+++ cmd/ausoceantv/frontend/src/app-lite/pages/home.html
@@ -1,0 +1,20 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/png" href="/src/assets/ausocean_logo.png" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>AusOcean TV | Home</title>
+    <link rel="stylesheet" href="./src/index.css" />
+    <script type="module" src="src/web-components/authenticator.ts"></script>
+  </head>
+  <body>
+    <auth-wrapper>
+      <div class="flex flex-col">
+        <p>AusOcean TV</p>
+        <a href="watch.html" class="mx-auto w-auto whitespace-nowrap rounded bg-gray-300 px-5 py-3 text-center">Watch</a>
+        <a href="/api/v1/auth/logout" class="mx-auto w-auto whitespace-nowrap rounded bg-gray-300 px-5 py-3 text-center">Logout</a>
+      </div>
+    </auth-wrapper>
+  </body>
+</html>
