+++ cmd/ausoceantv/frontend/src/app-lite/pages/index.html
@@ -1,0 +1,24 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/png" href="/src/assets/ausocean_logo.png" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>AusOcean TV | Welcome</title>
+    <link rel="stylesheet" href="./src/index.css" />
+    <script type="module" src="src/web-components/authenticator.ts"></script>
+  </head>
+  <body class="h-screen">
+    <auth-wrapper>
+      <div class="align-center container m-auto flex h-full flex-col justify-center">
+        <a href="https://ausocean.org" target="_blank">
+          <img src="src/assets/ausocean_logo.png" class="logo m-12 mx-auto w-64" alt="AusOcean logo" />
+        </a>
+        <h1 class="m-12 whitespace-nowrap text-center text-2xl font-bold sm:text-4xl md:text-6xl">Welcome to AusOcean TV</h1>
+        <button class="mx-auto w-auto whitespace-nowrap rounded bg-gray-300 px-5 py-3 text-center">
+          <a href="/api/v1/auth/login?redirect=/home.html">Sign Up</a>
+        </button>
+      </div>
+    </auth-wrapper>
+  </body>
+</html>
