+++ cmd/ausoceantv/frontend/src/shared/types/user.ts
