+++ cmd/ausoceantv/frontend/src/shared/web-components/global.d.ts
