+++ cmd/ausoceantv/frontend/tailwind.config.js
