+++ cmd/ausoceantv/frontend/tsconfig.json
