+++ cmd/ausoceantv/frontend/vite.config.ts
